do not generate TLS endpoints when the cp validation context is not present
