+++ pkg/xds/envoy/listeners/v3/server_mtls_with_cp_configurer.go
@@ -1,13 +1,10 @@
 package v3
 
 import (
-	semver "github.com/Masterminds/semver/v3"
 	envoy_core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 	envoy_listener "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
 	envoy_extensions_transport_sockets_tls_v3 "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
 
-	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	core_xds "github.com/kumahq/kuma/pkg/core/xds"
 	util_proto "github.com/kumahq/kuma/pkg/util/proto"
 	xds_context "github.com/kumahq/kuma/pkg/xds/context"
 	xds_tls "github.com/kumahq/kuma/pkg/xds/envoy/tls"
@@ -15,24 +12,10 @@ import (
 )
 
 type ServerSideMTLSWithCPConfigurer struct {
-	Ctx      xds_context.Context
-	Metadata *core_xds.DataplaneMetadata
+	Ctx xds_context.Context
 }
 
 var _ FilterChainConfigurer = &ServerSideMTLSWithCPConfigurer{}
-
-// backwards compatibility with 1.3.x
-var HasCPValidationCtxInBootstrap = func(version *mesh_proto.Version) (bool, error) {
-	if version.GetKumaDp().GetVersion() == "" { // mostly for tests but also for very old version of Kuma
-		return false, nil
-	}
-
-	semverVer, err := semver.NewVersion(version.KumaDp.GetVersion())
-	if err != nil {
-		return false, err
-	}
-	return !semverVer.LessThan(semver.MustParse("1.4.0")), nil
-}
 
 func (c *ServerSideMTLSWithCPConfigurer) Configure(filterChain *envoy_listener.FilterChain) error {
 	tlsContext, err := tls.CreateDownstreamTlsContext(c.Ctx)
@@ -43,22 +26,12 @@ func (c *ServerSideMTLSWithCPConfigurer) Configure(filterChain *envoy_listener.F
 		tlsContext = tls.StaticDownstreamTlsContext(c.Ctx.ControlPlane.AdminProxyKeyPair)
 	}
 
-	hasCpValidationCtx, err := HasCPValidationCtxInBootstrap(c.Metadata.GetVersion())
-	if err != nil {
-		return err
-	}
-
-	if hasCpValidationCtx {
-		// require certs that were presented when DP connects to the CP
-		tlsContext.RequireClientCertificate = util_proto.Bool(true)
-		tlsContext.CommonTlsContext.ValidationContextType = &envoy_extensions_transport_sockets_tls_v3.CommonTlsContext_ValidationContextSdsSecretConfig{
-			ValidationContextSdsSecretConfig: &envoy_extensions_transport_sockets_tls_v3.SdsSecretConfig{
-				Name: xds_tls.CpValidationCtx,
-			},
-		}
-	} else {
-		tlsContext.RequireClientCertificate = util_proto.Bool(false)
-		tlsContext.CommonTlsContext.ValidationContextType = nil
+	// require certs that were presented when DP connects to the CP
+	tlsContext.RequireClientCertificate = util_proto.Bool(true)
+	tlsContext.CommonTlsContext.ValidationContextType = &envoy_extensions_transport_sockets_tls_v3.CommonTlsContext_ValidationContextSdsSecretConfig{
+		ValidationContextSdsSecretConfig: &envoy_extensions_transport_sockets_tls_v3.SdsSecretConfig{
+			Name: xds_tls.CpValidationCtx,
+		},
 	}
 
 	pbst, err := util_proto.MarshalAnyDeterministic(tlsContext)
