+++ pkg/xds/generator/admin_proxy_generator.go
@@ -1,6 +1,9 @@
 package generator
 
 import (
+	"github.com/Masterminds/semver/v3"
+
+	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	core_xds "github.com/kumahq/kuma/pkg/core/xds"
 	xds_context "github.com/kumahq/kuma/pkg/xds/context"
 	envoy_common "github.com/kumahq/kuma/pkg/xds/envoy"
@@ -29,6 +32,19 @@ var staticTlsEndpointPaths = []*envoy_common.StaticEndpointPath{
 // AdminProxyGenerator generates resources to expose some endpoints of Admin API on public interface.
 // By default, Admin API is exposed only on loopback interface because of security reasons.
 type AdminProxyGenerator struct {
+}
+
+// backwards compatibility with 1.3.x
+var HasCPValidationCtxInBootstrap = func(version *mesh_proto.Version) (bool, error) {
+	if version.GetKumaDp().GetVersion() == "" { // mostly for tests but also for very old version of Kuma
+		return false, nil
+	}
+
+	semverVer, err := semver.NewVersion(version.KumaDp.GetVersion())
+	if err != nil {
+		return false, err
+	}
+	return !semverVer.LessThan(semver.MustParse("1.4.0")), nil
 }
 
 func (g AdminProxyGenerator) Generate(ctx xds_context.Context, proxy *core_xds.Proxy) (*core_xds.ResourceSet, error) {
@@ -66,14 +82,18 @@ func (g AdminProxyGenerator) Generate(ctx xds_context.Context, proxy *core_xds.P
 				Configure(envoy_listeners.StaticEndpoints(envoy_names.GetAdminListenerName(), staticEndpointPaths)),
 			),
 		}
-		if proxy.Dataplane != nil {
+		hasCpValidationCtx, err := HasCPValidationCtxInBootstrap(proxy.Metadata.Version)
+		if err != nil {
+			return nil, err
+		}
+		if proxy.Dataplane != nil && hasCpValidationCtx {
 			for _, se := range staticTlsEndpointPaths {
 				se.ClusterName = envoyAdminClusterName
 			}
 			filterChains = append(filterChains, envoy_listeners.FilterChain(envoy_listeners.NewFilterChainBuilder(proxy.APIVersion).
 				Configure(envoy_listeners.MatchTransportProtocol("tls")).
 				Configure(envoy_listeners.StaticEndpoints(envoy_names.GetAdminListenerName(), staticTlsEndpointPaths)).
-				Configure(envoy_listeners.ServerSideMTLSWithCP(ctx, proxy.Metadata)),
+				Configure(envoy_listeners.ServerSideMTLSWithCP(ctx)),
 			))
 		}
 		listener, err := envoy_listeners.NewListenerBuilder(proxy.APIVersion).
