+++ pkg/xds/generator/testdata/profile-source/2-envoy-config.golden.yaml
@@ -247,38 +247,6 @@ resources:
                   cluster: kuma:envoy:admin
                   prefixRewrite: /ready
           statPrefix: kuma_envoy_admin
-    - filterChainMatch:
-        transportProtocol: tls
-      filters:
-      - name: envoy.filters.network.http_connection_manager
-        typedConfig:
-          '@type': type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
-          httpFilters:
-          - name: envoy.filters.http.router
-          routeConfig:
-            validateClusters: false
-            virtualHosts:
-            - domains:
-              - '*'
-              name: kuma:envoy:admin
-              routes:
-              - match:
-                  prefix: /
-                route:
-                  cluster: kuma:envoy:admin
-                  prefixRewrite: /
-          statPrefix: kuma_envoy_admin
-      transportSocket:
-        name: envoy.transport_sockets.tls
-        typedConfig:
-          '@type': type.googleapis.com/envoy.extensions.transport_sockets.tls.v3.DownstreamTlsContext
-          commonTlsContext:
-            tlsCertificateSdsSecretConfigs:
-            - name: identity_cert
-              sdsConfig:
-                ads: {}
-                resourceApiVersion: V3
-          requireClientCertificate: false
     listenerFilters:
     - name: envoy.filters.listener.tls_inspector
       typedConfig:
